fix(product): resolve B02 pending-focus timing, keep both files under size cap
Reviewer re-review (2091-R2-1) found the prior B02 fix over-corrected: gating
setPendingFocus on the PRE-widen geometryPermits meant the production default
rail (420px -> 419px body) never focused the filter on toggle-open, and a
breadcrumb reveal from that geometry never revealed/focused its row -- the
spec's primary documented open paths. The spec's cancel rule is a POST-widen
outcome, not a pre-widen gate.

- use-file-editor-dock-controller.ts (new, extracted from FileEditorView.tsx):
  always mint the pending-focus token on toggle-open/breadcrumb-reveal. A new
  settle effect (keyed on bodyWidth/dockVisible) discards the token only if
  the requested geometry settles with the dock still not effectively visible,
  so a later unrelated responsive restore can never consume a stale token.
  If the widen lands with the dock visible, DockedFileTree consumes the token
  normally.
- FileEditorView.tsx / FileTreeDirectory.tsx: pure code motion (dock
  controller extracted to the hook above; FileTreeRetryRow extracted to its
  own colocated file) to bring both files under the repo's 400-line
  structure-report threshold with no behavior change.
- FileEditorView.file-tree-dock.test.tsx: replaced the over-corrected B02
  test with three tests that simulate the widen actually landing (via an
  `ensureRightPanelWidth` mock that resizes): first-open-from-default focuses
  the filter, a breadcrumb-open from default rail reveals/focuses its row,
  and a hard-capped-below-660 reveal discards its token so a later unrelated
  widen cannot steal focus.

report_frontend_structure.py --strict --summary-only now reports TOTAL: 0.
